Refactor message implementation into focused modules
